requirements.txt
addict==2.4.0
einops==0.8.0
matplotlib==3.8.4
monai==1.3.2
nibabel==5.3.2
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
scikit_learn==1.4.2
scipy==1.14.1
tensorflow==2.9.0
termcolor==2.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.40.2
yapf==0.40.2
